[7.1.0] Compute output directories in parallel when building the upload manifest. (#21386)
Some users have very large tree artifacts (with hundreds of thousands of
files) which are extremely slow to traverse sequentially. In one
particular scenario with ~250k files, this change reduces the time spent
computing the manifest from ~70s to ~5s.

(We could parallelize further across all outputs, but I prefer to wait
until there is a concrete need for it.)

PiperOrigin-RevId: 607678324
Change-Id: Ia80d0a663909b8ed96c6720d9eb5393fb3027b48