Remove unnecessary synchronization on `RepositoryCache`
With a single exception that didn't matter for correctness, the
operations on `RepositoryCache` were already atomic on the level of the
file system and thus don't require `synchronized`. In fact, since
different Bazel server instances share the same repository cache
concurrently, this already had to be the case.

Also removes unnecessary interrupt checks and method overloads only used
in tests.