[7.1.0] Share classpath `NestedSet` between full and header compile actions (#21389)
The Starlark Java rules already add the direct JARs to the transitive
classpath depset and pass the same depset to the full and the header
compile action. By allowing them to bypass the logic that prepends the
direct jars to the compile-time classpath, both actions retain the same
`NestedSet` instead of both retaining a new one with identical elements.

Closes #21343.

Commit
https://github.com/bazelbuild/bazel/commit/31fae9e8e6687cbaf0dfe55a466696210c80be96

PiperOrigin-RevId: 607790056
Change-Id: Ia08c834ae3e5b1151607459408cdfea85d47314f

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>