Replace a call to Spawn.getRunfilesSupplier() with getting the runfiles data from InputMetadataProvider.
This is seemingly not 100% functionally the same because the new code only iterates over tool inputs, but looking at the code, the runfiles middleman of a rule only ever ends up in the "hidden top level" output group (which I'll ignore although in theory weird Starlark code could give trouble) and in FilesToRunProvider, and the latter is in turn only ever used in SpawnAction.Builder.addTool() and the like.

RELNOTES: None.
PiperOrigin-RevId: 608273396
Change-Id: Ic9397141fe7e6e794b017c493032cd23f5056a31