Reuse runfiles in tests with --reuse_sandbox_directories
Test actions are guaranteed to have a runfiles directory with the test name as part of the name. The path to the directory is unique between tests. If two tests (foo and bar) have the directory <source-root>/pkg/my_runfiles as part of their runfiles and this directory contains 1000 files, we would be symlinking the 1000 files for each test since the paths do not coincide.

To make sure we can reuse the runfiles directory we must rename the old runfiles directory for the action that was stashed to the path that is expected by the current test.

PiperOrigin-RevId: 608273594
Change-Id: Iea46fa3600d29d9a5747c93f104c0a5b23d1ea53