Use cppConfiguration from cc_toolchain in cc_common.link
There's slight difference in semantics, because before the implementation sometimes used cppConfiguration from the current target and some rules use cc_toolchain from another rule.

The new semantics is better, because the configuration of cc_toolchain and what's being linked, matches.

It also helps Starlarkfying cc_common.link, where cppConfigration is not available (only actions are passed to the call and not the whole ctx).

PiperOrigin-RevId: 608516065
Change-Id: I1258f1d0afb8ff246b6623d1a07359517604e6d1