[7.1.0] Traverse symlinks to directories while collecting a TreeArtifactValue. (#21418)
As explained in https://github.com/bazelbuild/bazel/issues/20418, when a
tree artifact contains a symlink to a directory, it is collected as a
single TreeFileArtifact with DirectoryArtifactValue metadata. With this
change, the symlink is followed and the directory expanded into its
contents, which is more incrementally correct and removes a special case
that tree artifact consumers would otherwise have to be aware of.

This also addresses https://github.com/bazelbuild/bazel/issues/21171,
which is due to the metadata for the directory contents not being
available when building without the bytes, causing the input Merkle tree
builder to fail. (While I could have fixed this by falling back to
reading the directory contents from the filesystem, I prefer to abide by
the principle that input metadata should be collected before execution;
source directories are the other case where this isn't true, which I
also regard as a bug.)

Fixes https://github.com/bazelbuild/bazel/issues/20418.
Fixes https://github.com/bazelbuild/bazel/issues/21171.

Commit
https://github.com/bazelbuild/bazel/commit/4247c20f1daeeb787e9b5cb64848d398467effb1

PiperOrigin-RevId: 608389141
Change-Id: I956f3f8a4b1bfd279091e179d1cba3cdd0e5019b

Co-authored-by: Googler <tjgq@google.com>