[7.1.0] Correctly handle file inputs/outputs with directory contents in the execution log. (#21427)
https://github.com/bazelbuild/bazel/commit/9fd9aa57c456f5324383528c4094995f19124889
was a little too conservative: it didn't account for inputs/outputs
declared as a file but materialized as a directory, which are still
permitted by --noincompatible_disallow_unsound_directory_outputs.

Also fall back to the filesystem if input metadata isn't available to
make the determination. This shouldn't happen most of the time, as
metadata should already be available for most inputs, but there are some
hard-to-fix edge cases where it doesn't hold.

PiperOrigin-RevId: 608584354
Change-Id: If9a4c6df4319ef66f487068dd89b382eea1b425e