Handle the intersection of partially built targets and expanded outputs.
The set of built artifacts contains unexpanded artifacts. When filtering out lost outputs, we must map back to their owners. For example, when a tree child is lost, we should remove the parent tree artifact from the built artifacts set so that the lost URIs are not reported in the BEP.

PiperOrigin-RevId: 608633358
Change-Id: I8d6da4fc133f5e6c1a5b400aef67429bbab4a93c