Remove the restriction that relative symlinks in a tree artifact may not point outside the tree.
As discussed in https://github.com/bazelbuild/bazel/issues/20891, the restriction is pointless, as it can already be bypassed with an absolute symlink. Note that the symlinks are still required not to dangle in all cases.

Also rename and reorganize the tests in TreeArtifactBuildTest in a more logical manner.

Fixes https://github.com/bazelbuild/bazel/issues/20891.

Closes #21263.

PiperOrigin-RevId: 608926604
Change-Id: I967a383b9891360700f868abd5c2d292e0e7974e