Introduce a priority to select the child Exception in ErrorInfo.
Before: Arbitrarily select the first available Exception.

After: Priority should be given to the child Exceptions of keys that come without a value.

When both a value and an exception is present, that means the exception was tolerated along the way (in --keep_going mode). We should give priority to the more "meaningful" exceptions.

PiperOrigin-RevId: 608969186
Change-Id: I2e67beb28c2cacf74e577db0f4b00191402e6e32