Fix a conflict checking bug with Skymeld + memory saving mode.
In memory saving mode (signified by the various flags, among which is `--notrack_incremental_state`), conflict checking is done differently. With Skymeld, we accumulate the evaluated `ActionLookupKey`s and perform conflict checking on them whenever a top level target is ready for execution.

Simply checking for `ActionLookupKey`s turned out to be incomplete: we need to exclude `ActionTemplateExpansionKey` from this calculation. Those are only available during the execution phase and should not be considered in conflict checking, as it's WAI that `TreeFileArtifacts` would share a prefix conflict with their parent `TreeArtifact`.

This CL fixes that by excluding `ActionTemplateExpansionKey` from conflict checking.

PiperOrigin-RevId: 609397433
Change-Id: Id5ba8ebcf87de01baebe3eebc91ccd27dc85c6dd