Add prior actions executed and prior worker memory usage to worker metrics. This will allow us to calculate the actions executed / memory usage deltas as a result of that particular invocation.
This will be useful to help us find invocations that cause spikes in worker memory usage. The next step might be to figure out which actions cause the memory usage spikes, but then we'd need to log memory usage per-action in that case (while might not be feasible using ps, it might be possible using cgroups as that is much faster).

PiperOrigin-RevId: 609397589
Change-Id: I14b34597044e556fb80b723a008330ddfb35853e