Expand command lines into an `ImmutableList` instead of using the bespoke `IterablesChain`.
The only two callers (`SpawnAction` and `JavaCompileAction`) immediately make an `ImmutableList` copy anyway, so this saves garbage.

This also allows command lines to expand to any `Iterable` implementation, not just ones that satisfy `CollectionUtils#checkImmutable`. That check doesn't seem to be providing any value given that we immediately make an immutable copy anyway.

PiperOrigin-RevId: 609595064
Change-Id: I39f301270283772c8fb88ebec9d4855b901de4df