Add a method to `SingleStringArgFormatter` to get the would-be length of the formatted string without actually formatting.
This will be useful for lazy command line string formatting. We need to count the length of the command line to decide whether to spill the arguments to a param file, but we don't want to actually create the formatted strings.

In order to share logic with `format()`, extract a helper method that takes a null `StringBuilder` if no formatting is to be applied.

Also revamped the unit test to use test parameters.

PiperOrigin-RevId: 609829494
Change-Id: I5d640ca0b2f3a026dd09b02d1b5ef5c5c812c607