Fix stale trash dir not cleaned up on worker creation
The bug in Bazel is that ${output_base}/blaze-workers/_moved_trash_dir/ doesn't
get wiped between blaze restarts. The directory to move to for asynchronous
deletion is an incrementing AtomicInteger but after a restart it will be zero
again. If there were previous directories there, we might get an error trying
to replace a non-empty directory while renaming.

RELNOTES:none
PiperOrigin-RevId: 610323627
Change-Id: I2d397fbd4590fa9f83273ac11d92d591bf9348b8