[7.1.0] Fix genrule autostamping in bazel (#21512)
Genrule does not support conditional stamping based on the value of the
`--stamp` flag. This is because genrule treats the `stamp` attribute as
`boolean`, while all other rules use the `tristate` type. After this
change, the stamp attribute in genrule can be set to `-1` to request
conditional stamping.

RELNOTES: `genrule` now supports setting `stamp = -1` to request
conditional stamping (based on the value of the build's `--stamp` flag).

Closes https://github.com/bazelbuild/bazel/pull/21407.

Commit
https://github.com/bazelbuild/bazel/commit/cb901b71b4beb57b263bcb5551d0d45afc2e38c5,
https://github.com/bazelbuild/bazel/commit/ee57d991ccb56b78e8af0185f266e9a39554b1d4

PiperOrigin-RevId: 609085052
Change-Id: I873941e9aaae3760a545c1900cd13cb60a9ada39

---------

Co-authored-by: Googler <jhorvitz@google.com>
Co-authored-by: Alessandro Patti <ale812@yahoo.it>