Change the type of `InstrumentedFilesInfo.coverageEnvrionment` from a `NestedSet` to `ImmutableMap`
Using a `NestedSet` is unsuitable because on flattening, the value that finally ends up in the test environment depends on the structure of the set and ordering of children (in particular duplicate keys). Using a compile-order (aka post-order) `NestedSet` does not guarantee that the parent's entries will appear last because they may be de-duped by entries from a child.

 With a map, we can ensure that the parent environment "wins" by adding it to the map only after all it's children (true post-order).

PiperOrigin-RevId: 611061372
Change-Id: If545e723b6dd8e1dbefd72293bbaf532c1dc09e7