Allow filtering files to instrument for coverage
Bazel currently has a mechanism to skip targets from coverage
instrumentation via the --instrumentation_filter flag.  Unfortunately,
there are cases where individual _files_ within a target need to be
excluded.

This happens when a build contains gigantic targets with lots of source
files in them, some of which are problematic (e.g. because they contain
very large methods), and there is no way to easily build those files as
separate targets.

To resolve this, introduce --experimental_instrumentation_file_filter,
a new flag that applies filtering at the file level.

By looking at the code, it *seems* as if older versions of the
Java builder did have filtering as well, but such feature was lost
at some point in time.

Fixes #21520.