Splits SerializationContext into different subclass implementations.
This change splits out an explicitly ImmutableSerializationContext, which
can be retained in ObjectCodecs and freely shared across invocations.

It also separates MemoizingSerializationContext and
SharedValueSerializationContext (this could have alternatively
been named MemoizingAndBlockingSerializationContext, but
SharedValueSerializationContext seemed more purposeful) as their APIs
are different. The SharedValueSerializationContext must always return
a SerializationResult instead of just data due to the possibility of
serialization futures.

With this change, the construction of MemoizingSerializationContext
and SharedValueSerializationContext is fully encapsulated, with their
lifetimes clearly tied to the CodedOutputStreams they use. NestedSetStore
and HeaderInfoCodec can still instantiate these indirectly with calls
to getFreshContext but those will be cleaned up in later changes.

In preparation for a future change, SharedValueSerializationContext
needs to use a ByteArrayOutputStream instead of a ByteString.Output as
its underlying stream type because it needs to be able to edit the bytes
after the stream has been written. Consequently, there's an extra
ByteString.copyFrom. If this turns out to be a performance issue, we might
be able to set it reflectively at a later point in time.

PiperOrigin-RevId: 611179058
Change-Id: Ic79916fb326690abb2f3b4ff42a91de4cf780b0e