Add DUMPBIN make variable to MSVC C/C++ toolchain
`DUMPBIN` is a WIndows/MSVC tool that provides similar function to `NM` in the Linux world. `NM` already has a make variable, so I think it's fair to add one for `DUMPBIN`.

Tests and docs are updated as well.

RELNOTES[NEW]: New $(DUMPBIN) make variable is now available for Visual Studio toolchains.

Closes #21232.

PiperOrigin-RevId: 611398596
Change-Id: Ibe55d530c0dcace07372206741cc928f7c190dde