Fix flaky failures & timeouts in `StarlarkDebugIntegrationTest`
Trying to find an unused port and then attempting to use that for the debug server transport can occasionally fail, since we're essentially in a race to reclaim that port again. This is unlikely to happen locally but fairly common on a heavily loaded CI machine.

To avoid this, we instead bind to port 0 in the tests, and allow the tests to register a callback for the actual port returned by the kernel.

With the flaky failure and timeouts gone, we can also drop the sharding.

Before:

```
FAILED in 145 out of 1000 in 342.2s
  Stats over 1000 runs: max = 342.2s, min = 15.3s, avg = 62.3s, dev = 81.5s
```

After:

```
PASSED in 92.7s
  Stats over 1000 runs: max = 92.7s, min = 19.4s, avg = 34.0s, dev = 9.3s
```

PiperOrigin-RevId: 611407816
Change-Id: I0f6e4849619ebd458a2a9f6062e38b635f477b84