Reinstate "exactly one middleman" checks.
Now that scheduling middlemen are gone for good, they are only used for runfiles middlemen. Those should eventually be renamed to "runfiles artifacts", but for the time being, this is good enough.

The important part is that there were no checks that verified that the now-nonexistent scheduling middlemen exist.

RELNOTES: None.
PiperOrigin-RevId: 611444355
Change-Id: If75c4361c02c41e1d4553a9611a47b63658f4533