Post the first 5 action rewind events instead of the top 5.
A max of 5 action rewind events are posted as a sample, but we currently store every action rewind event in the whole build, then when the build ends, we sort them by number of lost inputs and take the top 5. In order to limit memory use when there is a lot of rewinding, just take the first 5 such events and do away with sorting by number of lost inputs. Note that we still post the total number of lost inputs in the build.

Removed assertions in a couple tests about exactly which events are posted since it can now be nondeterministic. I think these tests still offer excellent coverage.

PiperOrigin-RevId: 611566530
Change-Id: I190e6fc53d17945848fb2cfa7ae9bf6052450921