Defer string formatting in `StarlarkCustomCommandLine` by using lazy iterables.
Instead of eagerly building an `ImmutableList<String>`, command line expansion is split into two steps. The "preprocessing" step applies any exception-throwing operations such as directory expansion and `map_each`. String formatting operations such as `format_each`, `join_with`, and `flag_per_line` are applied during iteration over the command line.

The result is that parameter file arguments are not retained in memory throughout the lifetime of an action. Benchmarks show that for some real-life builds with many huge param files, this reduces peak heap size by 10G, old gen garbage by 97%, and cpu by 17.5%.

Note that if the command line does not get spilled to a param file, it is still materialized and held in memory for the duration of the action (`BaseSpawn#arguments`). We might be able to make this lazy too, but the benefits are unclear since anything large enough to cause blaze memory pressure should already be in a param file (in other words, command line limits help to keep this under control).

One additional challenge is that unless a param file is requested unconditionally (`Args.set_param_file_format(..., use_always = True)`), we need to know the total expanded command line length to test against `CommandLineLimits` and the param file threshold. This is implemented using calculations that derive the total length from the preprocessed representation without applying the string formatting.

Enhance unit tests for the implementation of various `Args` options in `StarlarkCustomCommandLine`. Although there is other test coverage for these features, this allows us to also verify that the new non-trival strategy for computing the total argument length is consistent with the old strategy of adding up string lengths.

PiperOrigin-RevId: 612505843
Change-Id: Ic979e34f829a288c0a9f42b27fb22b3329bd6467