skyfocus: include `DIRECTORY_LISTING_STATE` in verification sets.
Before this, the `FILE_STATE_VALUE` for directories were already captured in the working and verification sets, but that wasn't enough because changes to globs rely on `DIRECTORY_LISTING_STATE` invalidation. So this change keeps track of `DIRECTORY_LISTING_STATE` in the verification set, and prompts the user to add the directory to the working set if they're changed (whether via glob or other reasons).

See the test for the repro case.

RELNOTES:
PiperOrigin-RevId: 613458728
Change-Id: I7d70af00e566c82e851cf5a070fd7c155b039579