Implement new WorkerPoolImpl guarded by --use_new_worker_pool without using the apache.commons.pool2 library.
The implementation includes an additional mechanism to handle waiting borrowers when quota is unavailable (blocks) through the `waitingQueue` of `WorkerLatch(s)` and unblocks by signaling the latch in the event that a worker is returned or invalidated (w/o shrinking). While this mechanism is currently unused due to synchronization in the ResourceManager (as part of workers as a resource, causing it to only `#borrowObject` when resources are actually available), it is still included so for correctness sake, so that we don't need to assume that the API is called correctly.

PiperOrigin-RevId: 613525949
Change-Id: I2bb6508bdb144e0489a6453ca9ccf78af5332b66