skyfocus: ensure re-run if analysis cache drops (e.g. due to build setting change).
This adds a subscriber for `AnalysisPhaseCompleteEvent`, which contains information when the analysis cache was dropped (and therefore rebuilt), and needs a Skyfocus re-run. With Skymeld, `AnalysisPhaseCompleteEvent` is fired from a callback (when `AnalysisOperationWatcher#handleTopLevelEntityAnalysisConcluded` empties out the set of expected top level keys), so it's when analysis is _fully_ finished.

RELNOTES:
PiperOrigin-RevId: 613529397
Change-Id: I738c0e390cd2338dbbed909272dabe2aeb3cbe91