Add the path for in-memory jdeps files to execution info on demand instead of storing it in `JavaCompileAction` and `JavaHeaderCompileAction`.
Fixes memory overhead caused by `--experimental_inmemory_jdeps_files`.

The jdeps path added to the execution info map is novel for a given action, resulting in interning via `SpawnAction#executionInfoInterner` backfiring for `JavaHeaderCompileAction` - we get the interner overhead without the benefit, since entries are never equal. Instead, add the extra map entry when `getExecutionInfo()` is called. This approach is similar to how we implement `--experimental_inmemory_dotd_files`, see `CppCompileAction`.

This change means that `--modify_execution_info` can't strip the `internal-inline-outputs` key, which is actually a good thing given that it's internal, and is consistent with in-memory .d files.

PiperOrigin-RevId: 615419908
Change-Id: I48a5382007a5a3ebe246817a4723c9cef6314076