Ensure output root is an absolute path even if $HOME or $XDG_CACHE_HOME are relative
The Bazel server pre-processes the output root via PathFragmentConverter,
which absolutizes paths and expands leading "~" as the home directory. This
can cause the server and the client to treat different directories as output
root, leading to a server crash.

Thus, the client must make the output root absolute. We already do so when
the output root derives from $TEST_TMPDIR (without tilde expansion) in
StartupOptions::StartupOptions. And we do so (with tilde expansion, for
legacy reasons) when the output root derives from the --output_user_root
startup flag passed to the client, see StartupOptions::ProcessArg.

We must do the same when the output root derives from $HOME or
$XDG_CACHE_HOME - without tilde expansion, since it's generally expected
that tools do not perform tilde expansion on paths derived from env
variables (and since tilde expansion on $HOME is meaningless and a
recursive explosion).

Fixes https://github.com/bazelbuild/bazel/issues/21660

PiperOrigin-RevId: 615506594
Change-Id: Ice67c8be43cef2812b6dce591bd65b9813e82f5c