Stringify `Label`s in display form in `Args`
`Label`s added to `Args` are now formatted in display form, which means
that they will be repo-relative if referring to the main repo and use an
apparent repository name if possible.

Previously, labels were formatted in repo-relative form if referring to
the main repo and using canonical repository names exclusively for
external repos. At the same time, `Args` docs explicitly stated that
the exact stringification of any type other than `File` is unspecified.
The previous behavior was problematic since it neither always used
canonical labels (which could be useful for writing tests that check
these labels against an allowlist) nor provided labels that could be
included in BUILD files (canonical names are explicitly unstable).
Furthermore, whether a label resulted in a string prefixed with a single
or two `@`s already dependended on a user choice, namely the value of
`--enable_bzlmod`.

Thus, this change is not considered to be breaking. It makes it so that
existing rulesets relying on default stringification to return a BUILD-
file compatible label in WORKSPACE continue to do so with Bzlmod with no
code changes.

This change aims to be as memory efficient as possible:
* Single labels or sequences of labels that reference targets in the
  main repo incur no memory overhead.
* Labels referring to external repos as well as `NestedSet`s of labels
  result in an additional Skyframe dependency for each target using the
  command line as well as one additional reference (4 bytes) stored in
  the command line's `arguments`.