[7.1.1] Allow any canonical repo name to be used with `bazel mod show_repo` (#21694)
Previously we enforced that the canonical repo name had to correspond to
a module (i.e. not an extension-generated repo). This makes no sense, so
this change simple deletes that enforcement.

To compensate for the lack of checks whether the repo exists at all, we
add an additional check after requesting the BzlmodRepoRuleValues that
all requested repos must exist.

Fixes https://github.com/bazelbuild/bazel/issues/21621

Commit
https://github.com/bazelbuild/bazel/commit/5c2b37e56e77e9c96b1e5a6c99d803e31e17a03c

PiperOrigin-RevId: 615733963
Change-Id: Icf706b36f0e8d3c775bc26528948e5102910d4a0

Co-authored-by: Googler <wyv@google.com>