Stop documenting the distribs() function.
distribs() is non-functioning in Bazel and will be shortly removed from Blaze.

This was part of a larger CL that removes the function all at once. Since that is Build language impacting, it is difficult to line it all up at once w.r.t. depserver. I'm splitting out the parts so that the criticl CL which drops the behavior has a very small rollback surface.

RELNOTES: None
PiperOrigin-RevId: 616221367
Change-Id: I9672994a2e87432bcdfcddb337141f883f471ef3