Migrate SymbolicMacroTest.java to use text blocks
The test cases in this file are updated to use the text block feature available now that we've upgraded to JDK 21.

- Removed the empty "//" comments that were used as hints to the autoformatter to not mush all string arguments onto the same line.
- Replaced single quotes with double, since that's the preferred style in actual Starlark code. Single quoting was only used for readability to avoid escaping, which is no longer necessary.

In general, we shouldn't blindly migrate all tests to text blocks, since that obscures blame history. But in this case, the history of this file is relatively short and we're actively working on adding new test cases, so it seemed like a good tradeoff.

Work toward #19922.

PiperOrigin-RevId: 617153701
Change-Id: I1ecc70ecc47deb2386b2faf0ffd6d02621087e2f