[7.1.1] Fix sandbox cleanup crashing after server restart (#21733)
We try to clean up the sandbox base from previous server instances
asynchronously, however sometimes this is not possible due to the old
directories being in a different filesystem. This can happen with
overlays on
Docker after running bazel in different RUN commands. See
https://github.com/bazelbuild/bazel/issues/21719

This change fixes the crash by catching the IOException and falling back
to
synchronous deletion.

Fixes #21719.

RELNOTES:none
Commit
https://github.com/bazelbuild/bazel/commit/488b63072687899bdf7607bce949c47097fcd545

PiperOrigin-RevId: 617150522
Change-Id: I82a07ac0ade66cfb1e5732a90a5f3ab4e2e8caa7

Co-authored-by: Googler <plf@google.com>