Add a memprof_profile attribute to the fdo_profile rule.
This change extends the fdo_profile rule to optionally include a memprof_profile attribute. This attribute only supports a label and not an absolute path. The profile may be compressed (.zip) or uncompressed (.profdata). An additional configuration feature "memprof_optimize" is also needs to be present for the memprof attribute in the rule to be expanded to the expected compiler flags.

PiperOrigin-RevId: 617533321
Change-Id: Ib0e5d37022c9ce0835f0b1c273451ede7efed06c