Fix links to Protobufs on GitHub on the BEP docs page
A few links on [this page](https://bazel.build/remote/bep) are broken (the commit SHA for each is shortened, which apparently the GitHub UI doesn't accept). Since I was around, I thought it might have made sense to also update those pointers to a more recent version of the Protobufs in case someone uses it for reference.

The original pointers were supposed to go at the following locations

https://github.com/bazelbuild/bazel/blob/16a107dca10c49d2365886df4a06f50ce2e4aeb1/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto#L91
https://github.com/bazelbuild/bazel/blob/16a107dca10c49d2365886df4a06f50ce2e4aeb1/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto#L123
https://github.com/bazelbuild/bazel/blob/16a107dca10c49d2365886df4a06f50ce2e4aeb1/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto#L469

Closes #21745.

PiperOrigin-RevId: 618107617
Change-Id: Ib6794de510ebd850f3a3e22245bfcde92e53c825