Hard-wire --experimental_sibling_repository_layout to false.
According to the discussion
https://github.com/bazelbuild/bazel/discussions/20500, no one is too
attached to it and the project, sadly, failed: given the current
climate, it doesn't look like we'll ever have enough political will to
go through with the necessary migration to flip the default value of the
flag, as good an idea as it is on its own merits.

This change is designed to minimize the number of lines changed so that
it's easy to roll back if need be.