Adds Fingerprinter, to be used by Dumper to deduplicate by equivalence.
Introduces a Fingerprinter to assist in deduplication. The fingerprinter works
by performing a depth-first traversal of the object graph and composing
fingerprints bottom up.

One nuance here is how cycles are fingerprinted. To solve this in general,
something like canonicalization might work. However, that would be fairly
computationally expensive and complex. For now, the simpler alternative
of fingerprinting relative to the first-encountered node of the cycle
is taken. This means that fingerprint-based deduplication can't detect
isomorphic transformations.

Fingerprinter is extremely similar to Dumper, structurally, both being
traversals, but most of the code was too different to merge. The main
differences are as follows.

* Fingerprinter maintains a stack and needs to propagate information about
  cyclic backreferences via return values.
* Fingerprinter emits fingerprints when recursing, while Dumper increases the
  indentation level and emits inline.
* Dumper emits reference IDs, which Fingerprinter must not.

It might be possible to merge them by extracting an abstract traverser
that keeps information about the stack and cyclic references, but Dumper
never cares about those things, so it would make Dumper code more awkward,
especially the crucial methods involved in the core recursion.

Given that this is a test utility, favoring readability and simplicity
over code deduplication seems reasonable in this case.

PiperOrigin-RevId: 618887080
Change-Id: I1bd29b478dd409bf9d4cbb3b1dfa6cc468ffccd1