Split ModuleThreadContext off from ModuleFileGlobals
ModuleFileGlobals used to serve as both the definition of Starlark methods available in MODULE.bazel files, *and* the state tracker for a module being built. This commit brings it more in line with other Starlark evaluation contexts where there's a simple object that just holds methods, and the actual state is stored in the StarlarkThread. ModuleThreadContext is that state object.

This commit is a pure mechanical refactor with no behavioral changes. It paves the way for us to allow root module MODULE.bazel files to import other files.

Work towards https://github.com/bazelbuild/bazel/issues/17880.

Closes #21749.

PiperOrigin-RevId: 619261497
Change-Id: Ibff3a311e1788f7d8089ca194933cfb5fcc53ee0