Respect $XDG_CACHE_HOME if available for Bazel's output root on Linux and BSD
It may well be that $HOME is on a read-only mount. In this case, it is the
convention in the Linux and BSD world that $XDG_CACHE_HOME points to an
appropriate writable location.

Fixes https://github.com/bazelbuild/bazel/issues/16937

RELNOTES: Bazel on Linux and BSD now respects the XDG_CACHE_HOME environment
variable instead of assuming that ~/.cache/bazel is writable.
PiperOrigin-RevId: 614772057
Change-Id: I6377d7a90fb929843d18e82f5ed3d0adc55ac5c6