Refactor run-for-test into its own method
This makes it more clear how the test and non-test run paths do, and do not
overlap. Prior to this the sharing of variables and large if/else blocks gave
the impression that there was more sharing than there actually was.

Motivation is I'm trying to simplify --run_under, and this is a standalone
pre-factor which makes those changes/the possibility of them easier to
understand.

Includes a drive-by cleanup of unreachable windows-only escaping code in the
not-windows branch of transformArgvsForExecRequest.

PiperOrigin-RevId: 619271171
Change-Id: Ia41010d1f03d6cf3bee132c47f70bf58b54dc646