Extract --script_path-file generation into its own method
Script-path is its own special case, having it separate makes it a lot easier
to follow. There's an argument for the common location, but the indirection I
think outweighed the consolidation...

This is another standalone pre=factor as I sort out if/how run-under can be
simplified in the current implementation.

PiperOrigin-RevId: 619880362
Change-Id: I7a481be3f83dd3dc9c2c385e89fcadcdc5abb7fd