[7.2.0] Remove unnecessary synchronization on `RepositoryCache` (#21788)
With a single exception that didn't matter for correctness, the
operations on `RepositoryCache` were already atomic on the level of the
file system and thus don't require `synchronized`. In fact, since
different Bazel server instances share the same repository cache
concurrently, this already had to be the case.

The exception was a file write that stores the canonical id in a file
whose name contains a hash of the id. However, the content of this file
is never read, only its existence matters. The write is replaced by a
touch.

Also removes unnecessary interrupt checks and method overloads only used
in tests.

Work towards #20369

Closes #21403.

PiperOrigin-RevId: 614728666
Change-Id: I67f81bd9e468260e4f83f15d6aaafa57e34d18f4

Commit
https://github.com/bazelbuild/bazel/commit/8eade046a85670a940f7cf1a4f06f13d8b1a6518

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>