Add more logging to exception
We are seeing reports of an exception triggered due to unknown entries being
present in the sandbox base. Without actually knowing what those
files/directories are it is hard to know the next steps, for example if more
entries should be added to the allowlist SANDBOX_BASE_PERSISTENT_DIRS or if the
precondition check should be removed altogether.

This change adds more logging to get the names of the entries

RELNOTES:none
PiperOrigin-RevId: 621118864
Change-Id: Ie606f12a4148490d2c218ef5ba84605218e4f919