skyfocus: handle configuration changes and refactor SkyfocusModule state to use an immutable record stored in the SkyframeExecutor.
When the top level configuration changes, Blaze would detect and force an analysis cache drop. This means that a Skyfocused build will work as intended if one of the build flags change (e.g. --c opt -> fastbuild).

However, we can't rely on this behavior forever: if Blaze somehows figure out a way to only incrementally rebuild nodes that change, instead of discarding the full analysis cache, Skyfocus should still be aware and handle accordingly.

This CL introduces the --experimental_skyfocus_handling_strategy=<strict, warn>, which determines what Skyfocus should do when anything outside of the focused parts changed. For now, those are the verification set and the top level build configuration. The latter is handled by the new `SkyfocusState#checkConfigChanges` function in this CL.

It currently defaults to `strict`, so the build will fail if the top level build configuration value changes. Optionally, `warn` can be used to gracefully fallback to running the full-reanalysis, but not without acknowledging that the build configuration value has changed.

Additionally, this CL works towards dropping `@Subscribe` usage in `SkyfocusModule`.

PiperOrigin-RevId: 621412391
Change-Id: I15bd2b786ef02ac98e909cad4a0a23083081aac7