skyfocus: move callsite from `SkyfocusModule#onBuildToolFinalizingEvent` to `BuildTool#processRequest`, and remove `SkyfocusModule` as an EventBus listener.
This gets rid of the "effects at a distance", and makes reasoning about the Skyfocus lifecycle simpler.

The associated `BuildToolFinalizingEvent` is no longer needed, and `SkyfocusModule` is moved to the `runtime` package since it deals with runtime classes.

More specifically, any `AbruptExitException` thrown during Skyfocus will now be handled appropriately by the `BuildTool` crash management, and not actually crash Blaze server when it was called from an EventBus subscriber.

Also remove the `CommandEnvironment` member from `SkyfocusModule`.

PiperOrigin-RevId: 621427034
Change-Id: Id0de1a4a09b80bc8a9599043ca34aba1ff7e38c5