Implement "bazel dump --memory=shallow".
It allows one to display the objects retained by a particular SkyValue, but not its dependencies.

This is in contrast with --memory=deep, which counts the memory retained by a SkyValue, independently of whether any given object was created by itself or one of its transitive dependencies.

The implementation could be faster: it's currently single-threaded and traverses the direct dependencies sequentially instead of in parallel, but this will do now and is the simplest possible thing that could work.

RELNOTES: None.
PiperOrigin-RevId: 621792889
Change-Id: I874a4a27e5b5710de0a4d9e684496affcef6ae5e