Add `*_sdk_minimum_os` attributes to `xcode_version` and `XcodeProperties`
This allows rules to determine the minimum supported OS versions for a given Xcode version. One such use case is emulating Swift Package Manager’s `SupportedPlatform` spec.

This is a backwards compatible change. `xcode_version` accepts not setting these attributes, and will pass on `None` to the `XcodeProperties` constructor. The `XcodeProperties` constructor accepts not setting these attributes as well, defaulting to `None`. The `XcodeProperties` constructor converts `None` to be equal to `default_*_sdk_version`. And since when `default_*_sdk_version` is `None` a “reasonable default” is chosen, these new attributes will always have a value.

In the future we may want to use these values in `xcode_config` to set a better default `*_minimum_os`. That would be a breaking change though.

Signed-off-by: Brentley Jones <github@brentleyjones.com>